working out how the models will interact
