+++ astrophot/models/gaussian_model.py
@@ -305,8 +305,7 @@ class Gaussian_PSF(PSF_Model):
     @default_internal
     def evaluate_model(self, X=None, Y=None, image=None, parameters=None):
         if X is None:
-            Coords = image.get_coordinate_meshgrid()
-            X, Y = Coords - parameters["center"].value[..., None, None]
+            X, Y = image.get_coordinate_meshgrid()
         return self.radial_model(torch.sqrt(X ** 2 + Y ** 2), image, parameters)
