+++ astrophot/models/moffat_model.py
@@ -150,8 +150,7 @@ class Moffat_PSF(PSF_Model):
     
     def evaluate_model(self, X=None, Y=None, image=None, parameters=None):
         if X is None:
-            Coords = image.get_coordinate_meshgrid()
-            X, Y = Coords - parameters["center"].value[..., None, None]
+            X, Y = image.get_coordinate_meshgrid()
         return self.radial_model(
             self.radius_metric(X, Y, image=image, parameters=parameters),
             image=image,
