+++ astrophot/models/pixelated_psf_model.py
@@ -67,8 +67,7 @@ class Pixelated_PSF(PSF_Model):
     @default_internal
     def evaluate_model(self, X=None, Y=None, image=None, parameters=None, **kwargs):
         if X is None:
-            Coords = image.get_coordinate_meshgrid()
-            X, Y = Coords - parameters["center"].value[..., None, None]
+            X, Y = image.get_coordinate_meshgrid()
 
         # Convert coordinates into pixel locations in the psf image
         pX, pY = self.psf_model.plane_to_pixel(X, Y)
